+++ MatrixVector/MatrixSquare.cpp
@@ -240,8 +240,8 @@ static int Crout_LU_with_Pivoting_Solve(double *LU, double B[], const int pivot[
 
 bool MatrixSquare::Solve(std::span<const double> rhs, std::span<double> sol)
 {
-    Utilities::MyAssert(rhs.size() == m_dim);
-    Utilities::MyAssert(sol.size() == m_dim);
+    Utilities::MyAssert(static_cast<int>(rhs.size()) == m_dim);
+    Utilities::MyAssert(static_cast<int>(sol.size()) == m_dim);
     int status = 0;
     if (m_dirty)
     {
